CSS Classes & About Us
Further expanded, organized, and clarified the css classes. Ever pallete color, except 'main redwood' because it is the page background color, has a class for making a container that color. Each changes the font color too so it will be readable. Additionally, half the layout for the About Us finished. Functionally, it is very incomplete. I want just the elipticals showing originally, and upon clicking them, for the colorful container to become opaque and the below text visible. Right side of page is entirely unfinished but will eventually have 'Contact Us' section